Create DCMotor class
Create DCMotor class to control GPIO motors with the Raspberry Pi and
Wiring Pi library (accessed through the command line). To prevent
confusion all pin numbers are input using Broadcom numbering but are
then translated to Wiring Pi numbering via a hashmap. The class contains
methods to set the direction and motion of the motor to either running
forwards (1), running backwards (-1), or completely off (0). The method
is overloaded to allow the user to specify how long the motor should run
as well.